feat: rename policy doc files
