+++ solution/ui/regulations/css/scss/main.scss
@@ -45,5 +45,5 @@ $eregs-image-path: "../images" !default;
 @import "partials/supplemental_content";
 @import "partials/blocking_modal";
 @import "partials/iframe_container";
-@import "partials/policy_repository";
+@import "partials/subjects";
 @import 'print';
