+++ solution/ui/regulations/css/scss/partials/_subjects.scss
@@ -23,7 +23,7 @@
     }
 }
 
-#policyRepositoryApp.repository-view,
+#subjectsApp.repository-view,
 #policyRepositorySearchApp.repository-search-view {
     margin-bottom: 1rem;
     display: flex;
