+++ solution/ui/regulations/eregs-vite/src/components/subjects/SelectedSubjectHeading.vue
