+++ solution/ui/regulations/eregs-vite/src/components/subjects/SubjectChips.vue
